Add SideChainDonationReceived.
